apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.3
sqlalchemy==2.0.30
aiosqlite==0.20.0
stripe==10.0.0
resend==2.0.0
redis==5.0.6
fastapi-cache2==0.2.1
qdrant-client==1.9.1
sentence-transformers==3.0.1
transformers==4.42.4
peft==0.11.1
torch==2.3.1
bitsandbytes==0.43.1
accelerate==0.31.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
